requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
ImageIO==2.37.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
python-docx==1.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2026.3.3
xgboost==3.2.0
